refactor the System part of the daemon
